Move tests to batch_inputs_processor
